% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{adjacent_div}
\alias{adjacent_div}
\title{expand a number vector according to the adjacent two numbers}
\usage{
adjacent_div(v, n_div = 10, .unique = FALSE)
}
\arguments{
\item{v}{number vector}

\item{n_div}{how many divisions expanded by two numbers}

\item{.unique}{only keep unique numbers}
}
\value{
new number vector
}
\description{
expand a number vector according to the adjacent two numbers
}
\examples{
adjacent_div(10^c(1:3), n_div = 10)
}
