% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as_tibble_md}
\alias{as_tibble_md}
\title{trans a table in markdown format into tibble}
\usage{
as_tibble_md(x)
}
\arguments{
\item{x}{character string}
}
\value{
tibble
}
\description{
trans a table in markdown format into tibble
}
\examples{

x <- "
col1 | col2 | col3 |
| ---- | ---- | ---- |
| v1   | v2   | v3   |
| r1   | r2   | r3   |
"

as_tibble_md(x)

}
