% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baizer-package.R
\docType{package}
\name{baizer-package}
\alias{baizer}
\alias{baizer-package}
\title{baizer: Useful functions for data processing}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/william-swl/baizer}
  \item Report bugs at \url{https://github.com/william-swl/baizer/issues}
}

}
\author{
\strong{Maintainer}: William Song \email{william_swl@163.com}

}
\keyword{internal}
