% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{broadcast_vector}
\alias{broadcast_vector}
\title{broadcast the vector into length n}
\usage{
broadcast_vector(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{target length}
}
\value{
vector
}
\description{
broadcast the vector into length n
}
\examples{
broadcast_vector(1:3, 5)

}
