% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{c2r}
\alias{c2r}
\title{wrapper of tibble::column_to_rownames}
\usage{
c2r(df, col = "")
}
\arguments{
\item{df}{tibble}

\item{col}{a col name}
}
\value{
data.frame
}
\description{
wrapper of tibble::column_to_rownames
}
\examples{
mini_diamond \%>\% c2r("id")
}
