% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{check_arg}
\alias{check_arg}
\title{check arguments by custom function}
\usage{
check_arg(..., n = 2, fun = not.null)
}
\arguments{
\item{...}{arguments}

\item{n}{how many arguments should meet the custom conditions}

\item{fun}{custom conditions defined by a function}
}
\value{
logical value
}
\description{
check arguments by custom function
}
\examples{
x <- 1
y <- 3
z <- NULL

func <- function(x = NULL, y = NULL, z = NULL) {
  if (check_arg(x, y, z, n = 2)) {
    print("As expected, two arguments is not NULL")
  }

  if (check_arg(x, y, z, n = 1, method = ~ .x < 2)) {
    print("As expected, one argument less than 2")
  }
}

}
