% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{cmdargs}
\alias{cmdargs}
\title{get the command line arguments}
\usage{
cmdargs(x = NULL)
}
\arguments{
\item{x}{one of 'wd, R_env, script_path, script_dir, env_configs'}
}
\value{
list of all arguments, or single value of select argument
}
\description{
get the command line arguments
}
\examples{
cmdargs()

}
