% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{collapse_vector}
\alias{collapse_vector}
\title{dump a named vector into character}
\usage{
collapse_vector(named_vector, front_name = TRUE, collapse = ",")
}
\arguments{
\item{named_vector}{a named vector}

\item{front_name}{if TRUE, put names to former}

\item{collapse}{collapse separator}
}
\value{
character
}
\description{
dump a named vector into character
}
\examples{
collapse_vector(c(e = 1:4), front_name = TRUE, collapse = ";")
}
