% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{combn_vector}
\alias{combn_vector}
\title{combine multiple vectors into one}
\usage{
combn_vector(..., method = "first", invalid = NA)
}
\arguments{
\item{...}{vectors}

\item{method}{how to combine, should be one of
\code{first|last}, or one of \code{sum|mean|median} for numeric vector,
or some characters (e.g. \verb{,|.| |;}) for character vector}

\item{invalid}{invalid value to ignore, \code{NA} as default}
}
\value{
combined vector
}
\description{
combine multiple vectors into one
}
\examples{
x1 <- c(1, 2, NA, NA)
x2 <- c(3, NA, 2, NA)
x3 <- c(4, NA, NA, 3)

combn_vector(x1, x2, x3, method = "sum")

}
