% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{correct_ratio}
\alias{correct_ratio}
\title{correct the numbers to a target ratio}
\usage{
correct_ratio(raw, target, digits = 0)
}
\arguments{
\item{raw}{the raw numbers}

\item{target}{the target ratio}

\item{digits}{the result digits}
}
\value{
corrected number vector
}
\description{
correct the numbers to a target ratio
}
\examples{
correct_ratio(c(10, 10), c(3, 5))

# support ratio as a float
correct_ratio(c(100, 100), c(0.2, 0.8))

# more numbers
correct_ratio(10:13, c(2, 3, 4, 6))

# with digits after decimal point
correct_ratio(c(10, 10), c(1, 4), digits = 1)
}
