% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{cross_count}
\alias{cross_count}
\title{count two columns as a cross-tabulation table}
\usage{
cross_count(df, row, col, method = "n", digits = 2)
}
\arguments{
\item{df}{tibble}

\item{row}{the column as rownames in the output}

\item{col}{the column as colnames in the output}

\item{method}{one of \verb{n|count, rowr|row_ratio, colr|col_ratio}}

\item{digits}{the digits of ratios}
}
\value{
data.frame
}
\description{
count two columns as a cross-tabulation table
}
\examples{
cross_count(mini_diamond, cut, clarity)

# show the ratio in the row
cross_count(mini_diamond, cut, clarity, method = "rowr")

# show the ratio in the col
cross_count(mini_diamond, cut, clarity, method = "colr")

}
