% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{detect_dup}
\alias{detect_dup}
\title{detect possible duplication in a vector, ignore case,
blank and special character}
\usage{
detect_dup(vector, index = FALSE)
}
\arguments{
\item{vector}{vector possibly with duplication}

\item{index}{return duplication index}
}
\value{
duplication sub-vector
}
\description{
detect possible duplication in a vector, ignore case,
blank and special character
}
\examples{
detect_dup(c("a", "C_", "c -", "#A"))
}
