% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{diff_index}
\alias{diff_index}
\title{the index of different character}
\usage{
diff_index(s1, s2, nth = NULL, ignore_case = FALSE)
}
\arguments{
\item{s1}{string1}

\item{s2}{string2}

\item{nth}{just return nth index}

\item{ignore_case}{ignore upper or lower cases}
}
\value{
list of different character indices
}
\description{
the index of different character
}
\examples{
diff_index("AAAA", "ABBA")
}
