% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{diff_tb}
\alias{diff_tb}
\title{differences between two tibbles}
\usage{
diff_tb(old, new)
}
\arguments{
\item{old}{old tibble}

\item{new}{new tibble}
}
\value{
comparation tibble
}
\description{
differences between two tibbles
}
\examples{
tb1 <- gen_tb(fill = "int", seed = 1)

tb2 <- gen_tb(fill = "int", seed = 3)

diff_tb(tb1, tb2)

}
