% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{dx_tb}
\alias{dx_tb}
\title{diagnosis a tibble for character NA, NULL, all T/F column, blank in cell}
\usage{
dx_tb(x)
}
\arguments{
\item{x}{tibble}
}
\value{
list
}
\description{
diagnosis a tibble for character NA, NULL, all T/F column, blank in cell
}
\examples{
x <- tibble::tibble(
  c1 = c("NA", NA, "a", "b"),
  c2 = c("c", "d", "e", "NULL"),
  c3 = c("T", "F", "F", "T"),
  c4 = c("T", "F", "F", NA),
  c5 = c("", " ", "\t", "\n")
)

dx_tb(x)

}
