% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{exist_matrix}
\alias{exist_matrix}
\title{generate a matrix to show whether the item in each element of a list}
\usage{
exist_matrix(x, n_lim = 0, n_top = NULL, sort_items = NULL)
}
\arguments{
\item{x}{list of character vectors}

\item{n_lim}{n limit to keep items in result}

\item{n_top}{only keep top n items in result}

\item{sort_items}{function to sort the items, item frequency by default}
}
\value{
tibble
}
\description{
generate a matrix to show whether the item in each element of a list
}
\examples{
x <- 1:5 \%>\% purrr::map(
  ~ gen_char(to = "k", n = 5, random = TRUE, seed = .x)
)
exist_matrix(x)

}
