% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{expand_df}
\alias{expand_df}
\title{expand a dataframe by a value column}
\usage{
expand_df(df, name_col, value_col, sep = ",")
}
\arguments{
\item{df}{tibble}

\item{name_col}{repeat this as name column}

\item{value_col}{expand by this value column}

\item{sep}{separator in the string}
}
\value{
expanded tibble
}
\description{
expand a dataframe by a value column
}
\examples{
fancy_count(mini_diamond, "cut", "clarity") \%>\%
  expand_df(name_col = "cut", value_col = "clarity")
}
