% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{expr_pileup}
\alias{expr_pileup}
\title{pileup the subexpressions which is atomic}
\usage{
expr_pileup(ex)
}
\arguments{
\item{ex}{expression}
}
\value{
the character vector of subexpressions
}
\description{
pileup the subexpressions which is atomic
}
\examples{
ex <- rlang::expr(a == 2 & b == 3 | !b & x + 2)
expr_pileup(ex)

}
