% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{extract_kv}
\alias{extract_kv}
\title{extract key and values for a character vector}
\usage{
extract_kv(v, sep = ": ", key_loc = 1, value_loc = 2)
}
\arguments{
\item{v}{character vector}

\item{sep}{separator between key and value}

\item{key_loc}{key location}

\item{value_loc}{value location}
}
\value{
a named character vector
}
\description{
extract key and values for a character vector
}
\examples{
extract_kv(c("x: 1", "y: 2"))
}
