% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{fancy_count}
\alias{fancy_count}
\title{fancy count to show up to two columns and their summary}
\usage{
fancy_count(df, main_group, fine_group = NULL, fine_fmt = "count", sort = TRUE)
}
\arguments{
\item{df}{tibble}

\item{main_group}{first column as main group}

\item{fine_group}{second column as subgroup, can be ignore}

\item{fine_fmt}{output fine column format, \code{count|ratio|clean}}

\item{sort}{sort by frequency or not}
}
\value{
tibble
}
\description{
fancy count to show up to two columns and their summary
}
\examples{
fancy_count(mini_diamond, "cut", "clarity")
}
