% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{fetch_char}
\alias{fetch_char}
\title{fetch character from strings}
\usage{
fetch_char(s, index_list, na.rm = FALSE, collapse = FALSE)
}
\arguments{
\item{s}{strings}

\item{index_list}{index of nth character,
can be output of \code{diff_index} or \code{same_index}}

\item{na.rm}{remove NA values from results or not}

\item{collapse}{optional string used to combine
the characters from a same string}
}
\value{
list of characters
}
\description{
fetch character from strings
}
\examples{
fetch_char(rep("ABC", 3), list(1, 2, 3))
}
