% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbflt.R
\name{filterC}
\alias{filterC}
\title{apply tbflt on dplyr filter}
\usage{
filterC(.data, tbflt = NULL, .by = NULL)
}
\arguments{
\item{.data}{tibble}

\item{tbflt}{tbflt object}

\item{.by}{group by, same as \code{.by} argument in \code{dplyr::filter}}
}
\value{
tibble
}
\description{
apply tbflt on dplyr filter
}
\examples{
c1 <- tbflt(cut == "Fair")

c2 <- tbflt(x > 8)

mini_diamond \%>\%
  filterC(c1) \%>\%
  head(5)

mini_diamond \%>\% filterC(c1 & c2)

}
