% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{fix_to_regex}
\alias{fix_to_regex}
\title{trans fixed character into regex character}
\usage{
fix_to_regex(p)
}
\arguments{
\item{p}{raw fixed pattern}
}
\value{
regex pattern
}
\description{
trans fixed character into regex character
}
\examples{
fix_to_regex('ABC|?(*)')
}
