% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{float_to_percent}
\alias{float_to_percent}
\title{from float number to percent number}
\usage{
float_to_percent(x, digits = 2)
}
\arguments{
\item{x}{number}

\item{digits}{hold n digits after the decimal point}
}
\value{
percent character of x
}
\description{
from float number to percent number
}
\examples{
float_to_percent(0.12)
}
