% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{gen_char}
\alias{gen_char}
\title{generate characters}
\usage{
gen_char(
  from = NULL,
  to = NULL,
  n = NULL,
  random = FALSE,
  allow_dup = TRUE,
  add = NULL,
  seed = NULL
)
}
\arguments{
\item{from}{left bound, lower case letter}

\item{to}{right bound, lower case letter}

\item{n}{number of characters to generate}

\item{random}{random generation}

\item{allow_dup}{allow duplication when random generation}

\item{add}{add extra characters other than \code{base::letters}}

\item{seed}{random seed}
}
\value{
generated characters
}
\description{
generate characters
}
\examples{
gen_char(from = "g", n = 5)
gen_char(to = "g", n = 5)
gen_char(from = "g", to = "j")
gen_char(from = "t", n = 5, random = TRUE)
gen_char(
  from = "x", n = 5, random = TRUE,
  allow_dup = FALSE, add = c("+", "-")
)

}
