% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{gen_combn}
\alias{gen_combn}
\title{generate all combinations}
\usage{
gen_combn(x, n = 2)
}
\arguments{
\item{x}{vector}

\item{n}{numbers of element to combine}
}
\value{
all combinations
}
\description{
generate all combinations
}
\examples{
gen_combn(1:4, n = 2)

}
