% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{gen_outlier}
\alias{gen_outlier}
\title{generate outliers from a series of number}
\usage{
gen_outlier(x, n, digits = 0, side = "two.sided", lim = NULL, assign_n = NULL)
}
\arguments{
\item{x}{number vector}

\item{n}{number of outliers to generate}

\item{digits}{the digits of outliers}

\item{side}{should be one of \verb{two.sided, left, right}}

\item{lim}{a two-length vector to assign the limitations of the generation.
if method is \code{two.sided}, the generation will be limited in
[lim\link{1}, left-outlier-threshold] and [right-outlier-threshold, lim\link{2}];
if method is \code{left}, the generation will be limited in
[lim\link{1}, min(left-outlier-threshold, lim\link{2})]
if method is \code{right}, the generation will be limited in
[max(right-outlier-threshold, lim\link{1}), lim\link{2}]}

\item{assign_n}{manually assign the number of left outliers or
right outliers when method is \code{two.sided}}
}
\value{
number vector of outliers
}
\description{
generate outliers from a series of number
}
\examples{
x <- seq(0, 100, 1)

gen_outlier(x, 10)

# generation limits
gen_outlier(x, 10, lim = c(-80, 160))

# assign the left and right outliers
gen_outlier(x, 10, lim = c(-80, 160), assign_n = c(0.1, 0.9))

# just generate left outliers
gen_outlier(x, 10, side = "left")

}
