% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{gen_str}
\alias{gen_str}
\title{generate strings}
\usage{
gen_str(n = 1, len = 3, seed = NULL)
}
\arguments{
\item{n}{number of strings to generate}

\item{len}{string length}

\item{seed}{random seed}
}
\value{
string
}
\description{
generate strings
}
\examples{
gen_str(n = 2, len = 3)
}
