% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{gen_tb}
\alias{gen_tb}
\title{generate tibbles}
\usage{
gen_tb(nrow = 3, ncol = 4, fill = "float", colnames = NULL, seed = NULL, ...)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{fill}{fill by, one of \verb{float, int, char, str}}

\item{colnames}{names of columns}

\item{seed}{random seed}

\item{...}{parameters of \verb{rnorm, gen_char, gen_str}}
}
\value{
tibble
}
\description{
generate tibbles
}
\examples{
gen_tb()

gen_tb(fill = "str", nrow = 3, ncol = 4, len = 3)
}
