% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{generate_ticks}
\alias{generate_ticks}
\title{generate ticks for a number vector}
\usage{
generate_ticks(x, expect_ticks = 10)
}
\arguments{
\item{x}{number vector}

\item{expect_ticks}{expected number of ticks, may be a little different from
the result}
}
\value{
ticks number
}
\description{
generate ticks for a number vector
}
\examples{
generate_ticks(c(176, 198, 264))
}
