% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{geom_mean}
\alias{geom_mean}
\title{geometric mean}
\usage{
geom_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{value}

\item{na.rm}{remove NA or not}
}
\value{
geometric mean value
}
\description{
geometric mean
}
\examples{
geom_mean(1, 9)
}
