% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{\%eq\%}
\alias{\%eq\%}
\title{equal calculation operator, support NA}
\usage{
x \%eq\% y
}
\arguments{
\item{x}{value x}

\item{y}{value y}
}
\value{
logical value, TRUE if x and y are not equal
}
\description{
equal calculation operator, support NA
}
\examples{
NA \%eq\% NA
}
