% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{\%nin\%}
\alias{\%nin\%}
\title{\verb{\%nin\%}, a logical operator for not in calculation}
\usage{
left \%nin\% right
}
\arguments{
\item{left}{left element}

\item{right}{right element}
}
\value{
logical value, if left is not in right
}
\description{
\verb{\%nin\%}, a logical operator for not in calculation
}
\examples{
0 \%nin\% 1:4
}
