% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{\%!=na\%}
\alias{\%!=na\%}
\title{not equal calculation operator, support NA}
\usage{
x \%!=na\% y
}
\arguments{
\item{x}{value x}

\item{y}{value y}
}
\value{
logical value, TRUE if x and y are not equal
}
\description{
not equal calculation operator, support NA
}
\examples{
NA \%!=na\% NA
}
