% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{hist_bins}
\alias{hist_bins}
\title{separate numeric vector into bins}
\usage{
hist_bins(vector, bins = 10, sort = FALSE)
}
\arguments{
\item{vector}{numeric vector}

\item{bins}{bins number, defaults to 10}

\item{sort}{sort the result tibble}
}
\value{
tibble
}
\description{
separate numeric vector into bins
}
\examples{

vector <- dplyr::pull(mini_diamond, price, id)

hist_bins(vector, bins = 20)

}
