% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{int_digits}
\alias{int_digits}
\title{trans numbers to a fixed integer digit length}
\usage{
int_digits(x, digits = 2, scale_factor = FALSE)
}
\arguments{
\item{x}{number}

\item{digits}{integer digit length}

\item{scale_factor}{return the scale_factor instead of value}
}
\value{
number
}
\description{
trans numbers to a fixed integer digit length
}
\examples{
int_digits(0.0332, 1)
}
