% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{is.all_zero}
\alias{is.all_zero}
\title{if a number only have zeros}
\usage{
is.all_zero(x)
}
\arguments{
\item{x}{number}
}
\value{
all zero or not
}
\description{
if a number only have zeros
}
\examples{
is.all_zero('0.00')
}
