% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{is.blank}
\alias{is.blank}
\title{variable not in session calculation operator}
\usage{
is.blank(x)
}
\arguments{
\item{x}{variable}
}
\value{
logical value, TRUE if x not in session
}
\description{
variable not in session calculation operator
}
\examples{
is.blank(var_no_in_session)
}
