% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{is.exist}
\alias{is.exist}
\title{if variable exists}
\usage{
is.exist(x)
}
\arguments{
\item{x}{variable}
}
\value{
logical value, TRUE if x exists
}
\description{
if variable exists
}
\examples{
is.exist(var)
}
