% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{is.exsit}
\alias{is.exsit}
\title{if variable exsits}
\usage{
is.exsit(x, ls = NULL, envir = environment())
}
\arguments{
\item{x}{variable}

\item{ls}{variable list}

\item{envir}{environment}
}
\value{
logical value, TRUE if x exsits
}
\description{
if variable exsits
}
\examples{
is.exsit(var_not_in_session)
}
