% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{is.zero}
\alias{is.zero}
\title{if a number only have zeros}
\usage{
is.zero(x)
}
\arguments{
\item{x}{number}
}
\value{
all zero or not
}
\description{
if a number only have zeros
}
\examples{
is.zero(c("0.000", "0.102", NA))
}
