% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{left_expand}
\alias{left_expand}
\title{like \code{dplyr::left_join} while ignore the same columns in right tibble}
\usage{
left_expand(x, y, by = NULL)
}
\arguments{
\item{x}{left tibble}

\item{y}{right tibble}

\item{by}{columns to join by}
}
\value{
tibble
}
\description{
like \code{dplyr::left_join} while ignore the same columns in right tibble
}
\examples{

tb1 <- head(mini_diamond, 4)
tb2 <- tibble::tibble(
  id = c("id-2", "id-4", "id-5"),
  carat = 1:3,
  price = c(1000, 2000, 3000),
  newcol = c("new2", "new4", "new5")
)

left_expand(tb1, tb2, by = "id")

full_expand(tb1, tb2, by = "id")

inner_expand(tb1, tb2, by = "id")
}
