% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{list2df}
\alias{list2df}
\title{trans list into data.frame}
\usage{
list2df(x, rownames = TRUE, colnames = NULL, method = "row")
}
\arguments{
\item{x}{list}

\item{rownames}{use rownames or not}

\item{colnames}{colnames of the output}

\item{method}{one of \verb{row, col}, set each item as row or col, default as row}
}
\value{
tibble
}
\description{
trans list into data.frame
}
\examples{
x <- list(
  c("a", "1"),
  c("b", "2"),
  c("c", "3")
)

list2df(x, colnames = c("char", "num"))

x <- list(
  c("a", "b", "c"),
  c("1", "2", "3")
)

list2df(x, method = "col")
}
