% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{list2tibble}
\alias{list2tibble}
\title{tran list into tibble}
\usage{
list2tibble(x, colnames = NULL, method = "row")
}
\arguments{
\item{x}{list}

\item{colnames}{colnames of the output}

\item{method}{}
}
\value{
tibble
}
\description{
tran list into tibble
}
\examples{
x <- list(
    c('a', '1'),
    c('b', '2'),
    c('c', '3')
)

list2tibble(x, colnames=c('char', 'num'))

x <- list(
    c('a', 'b', 'c'),
    c('1', '2', '3')
)

list2tibble(x, method='col')
}
