% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{max_depth}
\alias{max_depth}
\title{max depth of a list}
\usage{
max_depth(x)
}
\arguments{
\item{x}{list}
}
\value{
number
}
\description{
max depth of a list
}
\examples{

max_depth(list(a = list(b = list(c = 1), d = 2, e = 3)))

}
