% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_diamond}
\alias{mini_diamond}
\title{Minimal tibble dataset adjusted from diamond}
\format{
\subsection{\code{mini_diamond}}{

A data frame with 100 rows and 7 columns:
\describe{
\item{id}{unique id}
\item{cut, clarity}{2 category variables}
\item{carat, price, x, y}{4 continuous variables}
...
}
}
}
\source{
adjusted from \code{ggplot2}
}
\usage{
mini_diamond
}
\description{
Minimal tibble dataset adjusted from diamond
}
\keyword{datasets}
