% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{mm_norm}
\alias{mm_norm}
\title{max-min normalization}
\usage{
mm_norm(x, low = 0, high = 1)
}
\arguments{
\item{x}{numeric vector}

\item{low}{low limit of result, 0 as default}

\item{high}{high limit of result, 1 as default}
}
\value{
normed vector
}
\description{
max-min normalization
}
\examples{

mm_norm(c(1, 3, 4))

}
