% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{move_row}
\alias{move_row}
\title{move selected rows to target location}
\usage{
move_row(df, rows, .after = FALSE, .before = FALSE)
}
\arguments{
\item{df}{tibble}

\item{rows}{selected rows indexes}

\item{.after}{\code{TRUE} will move selected rows to the last row,
or you can pass a target row index}

\item{.before}{\code{TRUE} will move selected rows to the first row,
or you can pass a target row index}
}
\value{
reordered tibble
}
\description{
move selected rows to target location
}
\examples{
move_row(mini_diamond, 3:5, .after = 8)
}
