% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{near_ticks}
\alias{near_ticks}
\title{the ticks near a number}
\usage{
near_ticks(x, level = NULL, div = 2)
}
\arguments{
\item{x}{number}

\item{level}{the level of ticks, such as 1, 10, 100, etc.}

\item{div}{number of divisions}
}
\value{
number vector of ticks
}
\description{
the ticks near a number
}
\examples{
near_ticks(3462, level = 10)
}
