% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{number_fun_wrapper}
\alias{number_fun_wrapper}
\title{wrapper of the functions to process number string with prefix and suffix}
\usage{
number_fun_wrapper(
  x,
  fun = ~.x,
  prefix_ext = NULL,
  suffix_ext = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{number string vector with prefix and suffix}

\item{fun}{process function}

\item{prefix_ext}{prefix extension}

\item{suffix_ext}{suffix extension}

\item{verbose}{print more details}
}
\value{
processed number with prefix and suffix
}
\description{
wrapper of the functions to process number string with prefix and suffix
}
\examples{
number_fun_wrapper(">=2.134\%", function(x) round(x, 2))
}
