% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{ordered_slice}
\alias{ordered_slice}
\title{slice a tibble by an ordered vector}
\usage{
ordered_slice(df, by, ordered_vector, na.rm = FALSE, dup.rm = FALSE)
}
\arguments{
\item{df}{tibble}

\item{by}{slice by this column, this value must has no duplicated value}

\item{ordered_vector}{ordered vector}

\item{na.rm}{remove NA or unknown values from ordered vector}

\item{dup.rm}{remove duplication values from ordered vector}
}
\value{
sliced tibble
}
\description{
slice a tibble by an ordered vector
}
\examples{
ordered_slice(mini_diamond, id, c("id-3", "id-2"))
}
