% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{pileup_logical}
\alias{pileup_logical}
\title{pileup another logical vector on the TRUE values of first vector}
\usage{
pileup_logical(x, v)
}
\arguments{
\item{x}{logical vector}

\item{v}{another logical vector}
}
\value{
logical vector
}
\description{
pileup another logical vector on the TRUE values of first vector
}
\examples{

# first vector have 2 TRUE value
v1 <- c(TRUE, FALSE, TRUE)

# the length of second vector should also be 2
v2 <- c(FALSE, TRUE)

pileup_logical(v1, v2)

}
