% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{pkginfo}
\alias{pkginfo}
\title{information of packages}
\usage{
pkginfo(...)
}
\arguments{
\item{...}{case-insensitive package names}
}
\description{
information of packages
}
\examples{
baizer::pkginfo(dplyr)
}
