% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{pkglib}
\alias{pkglib}
\title{load packages as a batch}
\usage{
pkglib(...)
}
\arguments{
\item{...}{pkgs}
}
\value{
nothing
}
\description{
load packages as a batch
}
\examples{
baizer::pkglib(dplyr, purrr, tidyr)
}
