% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{pkgver}
\alias{pkgver}
\title{versions of packages}
\usage{
pkgver(...)
}
\arguments{
\item{...}{case-insensitive package names}
}
\description{
versions of packages
}
\examples{
baizer::pkgver(dplyr, purrr)
}
