% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{pos_int_split}
\alias{pos_int_split}
\title{split a positive integer number as a number vector}
\usage{
pos_int_split(x, n, method = "average")
}
\arguments{
\item{x}{positive integer}

\item{n}{length of the output}

\item{method}{should be one of \verb{average, random}, or a number vector which
length is n}
}
\value{
number vector
}
\description{
split a positive integer number as a number vector
}
\examples{
pos_int_split(12, 3, method = "average")

pos_int_split(12, 3, method = "random")

pos_int_split(12, 3, method = c(1, 2, 3))

}
