% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{r2c}
\alias{r2c}
\title{wrapper of tibble::rownames_to_column}
\usage{
r2c(df, col = "")
}
\arguments{
\item{df}{tibble}

\item{col}{a col name}
}
\value{
tibble
}
\description{
wrapper of tibble::rownames_to_column
}
\examples{
mini_diamond \%>\%
  c2r("id") \%>\%
  r2c("id")
}
