% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_excel_list}
\alias{read_excel_list}
\title{read multi-sheet excel file as a list of tibbles}
\usage{
read_excel_list(x)
}
\arguments{
\item{x}{path}
}
\value{
list
}
\description{
read multi-sheet excel file as a list of tibbles
}
