% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_fmmd}
\alias{read_fmmd}
\title{read front matter markdown}
\usage{
read_fmmd(x, rm_blank_line = TRUE)
}
\arguments{
\item{x}{path}

\item{rm_blank_line}{remove leading and trailing blank lines}
}
\value{
list
}
\description{
read front matter markdown
}
