% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{ref_level}
\alias{ref_level}
\title{relevel a target column by another reference column}
\usage{
ref_level(x, col, ref)
}
\arguments{
\item{x}{tibble}

\item{col}{target column}

\item{ref}{reference column}
}
\value{
tibble
}
\description{
relevel a target column by another reference column
}
\examples{

cut_level <- mini_diamond \%>\%
  dplyr::pull(cut) \%>\%
  unique()

mini_diamond \%>\%
  dplyr::mutate(cut = factor(cut, cut_level)) \%>\%
  dplyr::mutate(cut0 = stringr::str_c(cut, "xxx")) \%>\%
  ref_level(cut0, cut)
}
