% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{reg_join}
\alias{reg_join}
\title{join the matched parts into string}
\usage{
reg_join(x, pattern, sep = "")
}
\arguments{
\item{x}{character}

\item{pattern}{regex pattern}

\item{sep}{separator}
}
\value{
character
}
\description{
join the matched parts into string
}
\examples{
reg_join(c("A_12.B", "C_3.23:2"), "[A-Za-z]+")

reg_join(c("A_12.B", "C_3.23:2"), "\\\\w+")

reg_join(c("A_12.B", "C_3.23:2"), "\\\\d+", sep = ",")

reg_join(c("A_12.B", "C_3.23:2"), "\\\\d", sep = ",")

}
