% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{reg_match}
\alias{reg_match}
\title{regex match}
\usage{
reg_match(x, pattern, group = 1)
}
\arguments{
\item{x}{vector}

\item{pattern}{regex pattern}

\item{group}{regex gruop, 1 as default. when group=-1,
return full matched tibble}
}
\value{
vector or tibble
}
\description{
regex match
}
\examples{
v <- stringr::str_c("id", 1:3, c("A", "B", "C"))

reg_match(v, "id(\\\\d+)(\\\\w)")

reg_match(v, "id(\\\\d+)(\\\\w)", group = 2)

reg_match(v, "id(\\\\d+)(\\\\w)", group = -1)

}
