% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{remove_monocol}
\alias{remove_monocol}
\title{remove columns by the ratio of an identical single value (NA supported)}
\usage{
remove_monocol(df, max_ratio = 1)
}
\arguments{
\item{df}{tibble}

\item{max_ratio}{the max single value ratio to keep this column, default is 1}
}
\value{
tibble
}
\description{
remove columns by the ratio of an identical single value (NA supported)
}
\examples{
# remove_monocol(df)
}
