% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{remove_nacol}
\alias{remove_nacol}
\title{remove columns only have NA value}
\usage{
remove_nacol(df)
}
\arguments{
\item{df}{tibble}
}
\value{
tibble
}
\description{
remove columns only have NA value
}
\examples{
# remove_nacol(df)
}
