% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{remove_narow}
\alias{remove_narow}
\title{remove rows by the ratio of NA}
\usage{
remove_narow(df, ..., max_ratio = 1)
}
\arguments{
\item{df}{tibble}

\item{...}{only remove rows according to these columns,
refer to \code{dplyr::select()}}

\item{max_ratio}{the max NA ratio to keep this row, default is 1
have NA}
}
\value{
tibble
}
\description{
remove rows by the ratio of NA
}
\examples{
# remove_narow(df)
}
