% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove outliers and NA}
\usage{
remove_outliers(df, col, .by = NULL)
}
\arguments{
\item{df}{tibble}

\item{col}{columns to remove outliers}

\item{.by}{group by}
}
\value{
tibble
}
\description{
remove outliers and NA
}
\examples{
remove_outliers(mini_diamond, price)
}
