% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{replace_item}
\alias{replace_item}
\title{replace the items of one object by another}
\usage{
replace_item(x, y, keep_extra = FALSE)
}
\arguments{
\item{x}{number, character or list}

\item{y}{another object, the class of y should be same as x}

\item{keep_extra}{whether keep extra items in y}
}
\value{
replaced object
}
\description{
replace the items of one object by another
}
\examples{

x <- list(A = 1, B = 3)
y <- list(A = 9, C = 10)

replace_item(x, y)

replace_item(x, y, keep_extra = TRUE)

}
