% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{rewrite_na}
\alias{rewrite_na}
\title{rewrite the NA values in a tibble by another tibble}
\usage{
rewrite_na(x, y, by)
}
\arguments{
\item{x}{raw tibble}

\item{y}{replace reference tibble}

\item{by}{columns to align the tibbles}
}
\value{
tibble
}
\description{
rewrite the NA values in a tibble by another tibble
}
\examples{

tb1 <- tibble::tibble(
  id = c("id-1", "id-2", "id-3", "id-4"),
  group = c("a", "b", "a", "b"),
  price = c(0, -200, 3000, NA),
  type = c("large", "none", "small", "none")
)

tb2 <- tibble::tibble(
  id = c("id-1", "id-2", "id-3", "id-4"),
  group = c("a", "b", "a", "b"),
  price = c(1, 2, 3, 4),
  type = c("l", "x", "x", "m")
)

rewrite_na(tb1, tb2, by = c("id", "group"))
}
