% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{rng2seq}
\alias{rng2seq}
\title{trans range character into seq characters}
\usage{
rng2seq(x, sep = "-")
}
\arguments{
\item{x}{range character}

\item{sep}{range separator}
}
\value{
seq characters
}
\description{
trans range character into seq characters
}
\examples{
rng2seq(c("1-5", "2"))
}
