% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{round_string}
\alias{round_string}
\title{from float number to fixed digits character}
\usage{
round_string(x, digits = 2)
}
\arguments{
\item{x}{number}

\item{digits}{hold n digits after the decimal point}
}
\value{
character
}
\description{
from float number to fixed digits character
}
\examples{
round_string(1.1, 2)
}
