% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{roxygen_fmt}
\alias{roxygen_fmt}
\title{add #' into each line of codes for roxygen examples}
\usage{
roxygen_fmt(x)
}
\arguments{
\item{x}{codes}
}
\description{
add #' into each line of codes for roxygen examples
}
\examples{

roxygen_fmt(
  "
code line1
code line2
"
)

}
