% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{same_index}
\alias{same_index}
\title{the index of identical character}
\usage{
same_index(s1, s2, nth = NULL, ignore_case = FALSE)
}
\arguments{
\item{s1}{string1}

\item{s2}{string2}

\item{nth}{just return nth index}

\item{ignore_case}{ignore upper or lower cases}
}
\value{
list of identical character indices
}
\description{
the index of identical character
}
\examples{
same_index("AAAA", "ABBA")
}
