% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{seriate_df}
\alias{seriate_df}
\title{dataframe rows seriation, which will reorder the rows in a better pattern}
\usage{
seriate_df(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
seriated dataframe
}
\description{
dataframe rows seriation, which will reorder the rows in a better pattern
}
\examples{
x <- mini_diamond \%>\%
  dplyr::select(id, dplyr::where(is.numeric)) \%>\%
  dplyr::mutate(
    dplyr::across(
      dplyr::where(is.numeric),
      ~ round(.x / max(.x), 4)
    )
  ) \%>\%
  c2r("id")

seriate_df(x)

}
