% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{sftp_connect}
\alias{sftp_connect}
\title{connection parameters to remote server via sftp}
\usage{
sftp_connect(
  server = "localhost",
  port = 22,
  user = NULL,
  password = NULL,
  wd = "~"
)
}
\arguments{
\item{server}{remote server}

\item{port}{SSH port, 22 as default}

\item{user}{username}

\item{password}{password}

\item{wd}{workdir}
}
\value{
sftp_connection object
}
\description{
connection parameters to remote server via sftp
}
\examples{
# sftp_con <- sftp_connect(server='remote_host', port=22,
#     user='username', password = "password", wd='~')
}
