% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{sftp_download}
\alias{sftp_download}
\title{download file from remote server via sftp}
\usage{
sftp_download(sftp_con, path = NULL, to = basename(path))
}
\arguments{
\item{sftp_con}{sftp_connection created by sftp_connect()}

\item{path}{remote file path}

\item{to}{local target path}
}
\description{
download file from remote server via sftp
}
\examples{
# sftp_download(sftp_con,
#   path=c('t1.txt', 't2.txt'),
#   to=c('path1.txt', 'path2.txt')
}
