% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{sftp_ls}
\alias{sftp_ls}
\title{list files from remote server via sftp}
\usage{
sftp_ls(sftp_con, path = NULL, all = FALSE)
}
\arguments{
\item{sftp_con}{sftp_connection created by sftp_connect()}

\item{path}{remote directory path}

\item{all}{list hidden files or not}
}
\value{
files in the dir
}
\description{
list files from remote server via sftp
}
\examples{

# sftp_ls(sftp_con, 'your/dir')

}
