% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{signif_ceiling}
\alias{signif_ceiling}
\title{signif while use ceiling}
\usage{
signif_ceiling(x, digits = 2)
}
\arguments{
\item{x}{number}

\item{digits}{digits}
}
\value{
number
}
\description{
signif while use ceiling
}
\examples{
signif_ceiling(3.11, 2)
}
