% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{signif_floor}
\alias{signif_floor}
\title{signif while use floor}
\usage{
signif_floor(x, digits = 2)
}
\arguments{
\item{x}{number}

\item{digits}{digits}
}
\value{
number
}
\description{
signif while use floor
}
\examples{
signif_floor(3.19, 2)
}
