% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{signif_round_string}
\alias{signif_round_string}
\title{signif or round string depend on the character length}
\usage{
signif_round_string(
  x,
  digits = 2,
  format = "short",
  full_large = TRUE,
  full_small = FALSE
)
}
\arguments{
\item{x}{number}

\item{digits}{signif or round digits}

\item{format}{short or long}

\item{full_large}{keep full digits for large number}

\item{full_small}{keep full digits for small number}
}
\value{
signif or round strings
}
\description{
signif or round string depend on the character length
}
\examples{
signif_round_string(1.214, 2)
}
