% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{signif_string}
\alias{signif_string}
\title{from float number to fixed significant digits character}
\usage{
signif_string(x, digits = 2)
}
\arguments{
\item{x}{number}

\item{digits}{hold n significant digits}
}
\value{
character
}
\description{
from float number to fixed significant digits character
}
\examples{
signif_string(1.1, 2)
}
