% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{slice_char}
\alias{slice_char}
\title{slice character vector}
\usage{
slice_char(x, from = x[1], to = x[length(x)], unique = FALSE)
}
\arguments{
\item{x}{character vector}

\item{from}{from}

\item{to}{to}

\item{unique}{remove the duplicated boundary characters}
}
\value{
sliced vector
}
\description{
slice character vector
}
\examples{
x <- c("A", "B", "C", "D", "E")
slice_char(x, "A", "D")
slice_char(x, "D", "A")

x <- c("A", "B", "C", "C", "A", "D", "D", "E", "A")
slice_char(x, "B", "E")
# duplicated element as boundary will throw an error
# slice_char(x, 'A', 'E')
# unique=TRUE to remove the duplicated boundary characters
slice_char(x, "A", "E", unique = TRUE)

}
