% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{sortf}
\alias{sortf}
\title{sort by a function}
\usage{
sortf(x, func, order = FALSE)
}
\arguments{
\item{x}{vector}

\item{func}{a function used by the sort}

\item{order}{only return the order}
}
\value{
vector
}
\description{
sort by a function
}
\examples{
v <- stringr::str_c('id', c(1, 2, 9, 10, 11, 12, 99, 101, 102)) \%>\% sort

sortf(v, function(x) reg_match(x, '\\\\d+') \%>\% as.double)

sortf(v, ~reg_match(.x, '\\\\d+') \%>\% as.double)

sortf(v, ~reg_match(.x, '\\\\d+') \%>\% as.double, order=TRUE)

}
