% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{split_column}
\alias{split_column}
\title{split a column and return a longer tibble}
\usage{
split_column(df, name_col, value_col, sep = ",")
}
\arguments{
\item{df}{tibble}

\item{name_col}{repeat this as name column}

\item{value_col}{expand by this value column}

\item{sep}{separator in the string}
}
\value{
expanded tibble
}
\description{
split a column and return a longer tibble
}
\examples{
fancy_count(mini_diamond, cut, ext = clarity) \%>\%
  split_column(name_col = cut, value_col = clarity)
}
