% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{split_path}
\alias{split_path}
\title{split a path into ancestor paths recursively}
\usage{
split_path(path)
}
\arguments{
\item{path}{path to split}
}
\value{
character vectors of ancestor paths
}
\description{
split a path into ancestor paths recursively
}
\examples{
split_path("/home/someone/a/test/path.txt")
}
