% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{split_vector}
\alias{split_vector}
\title{split vector into list}
\usage{
split_vector(vector, breaks, bounds = "(]")
}
\arguments{
\item{vector}{vector}

\item{breaks}{split breaks}

\item{bounds}{"(]" as default, can also be "[), []"}
}
\value{
list
}
\description{
split vector into list
}
\examples{
split_vector(1:10, c(3, 7))
split_vector(stringr::str_split("ABCDEFGHIJ", "") \%>\% unlist(),
  c(3, 7),
  bounds = "[)"
)
}
