% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_fc}
\alias{stat_fc}
\title{fold change calculation which returns a extensible tibble}
\usage{
stat_fc(df, y, x, method = "mean", .by = NULL, signif_digits = 2)
}
\arguments{
\item{df}{tibble}

\item{y}{value}

\item{x}{sample test group}

\item{method}{\code{'mean'|'median'|'geom_mean'}, the summary method}

\item{.by}{super-group}

\item{signif_digits}{fold change signif digits}
}
\value{
fold change result tibble
}
\description{
fold change calculation which returns a extensible tibble
}
\examples{
stat_fc(mini_diamond, y = price, x = cut, .by = clarity)
}
