% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_phi}
\alias{stat_phi}
\title{calculate phi coefficient of two binary variables}
\usage{
stat_phi(x)
}
\arguments{
\item{x}{2x2 matrix or dataframe}
}
\value{
phi coefficient
}
\description{
calculate phi coefficient of two binary variables
}
\examples{
data <- matrix(c(10, 8, 14, 18), nrow = 2)
stat_phi(data)
}
