% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{str_replace_loc}
\alias{str_replace_loc}
\title{replace specific characters in a string by their locations}
\usage{
str_replace_loc(x, start = 1, end = nchar(x), replacement = " ")
}
\arguments{
\item{x}{string}

\item{start}{start}

\item{end}{end}

\item{replacement}{replacement}
}
\value{
replaced string
}
\description{
replace specific characters in a string by their locations
}
\examples{
str_replace_loc("abcde", 1, 3, "A")
}
