% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbflt.R
\name{tbflt}
\alias{tbflt}
\title{create a tbflt object to save filter conditions}
\usage{
tbflt(x = expression())
}
\arguments{
\item{x}{any expression}
}
\value{
tbflt
}
\description{
\code{tbflt()} can save a series of filter conditions, and support
logical operating among conditions
}
\examples{
c1 <- tbflt(cut == "Fair")

c2 <- tbflt(x > 8)

!c1

c1 | c2

c1 & c2

}
