% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{tdf}
\alias{tdf}
\title{transpose a dataframe}
\usage{
tdf(x, colnames = NULL)
}
\arguments{
\item{x}{dataframe}

\item{colnames}{column names of the transposed dataframe}
}
\value{
dataframe
}
\description{
transpose a dataframe
}
\examples{

x <- c2r(mini_diamond, "id")
tdf(x)

}
