% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{top_item}
\alias{top_item}
\title{return top n items with highest frequency}
\usage{
top_item(x, n = 1)
}
\arguments{
\item{x}{character}

\item{n}{top n}
}
\value{
character
}
\description{
return top n items with highest frequency
}
\examples{

top_item(c("a", "b", "c", "b"))

}
