% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{uniq}
\alias{uniq}
\title{only keep unique vector values and its names}
\usage{
uniq(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
only keep unique vector values and its names
}
\examples{

x <- c(a = 1, b = 2, c = 3, b = 2, a = 1)

uniq(x)

}
