% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{uniq_in_cols}
\alias{uniq_in_cols}
\title{count unique values in each column}
\usage{
uniq_in_cols(x)
}
\arguments{
\item{x}{tibble}
}
\value{
tibble
}
\description{
count unique values in each column
}
\examples{

uniq_in_cols(mini_diamond)

}
