% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{vector_dump}
\alias{vector_dump}
\title{dump a named vector into character}
\usage{
vector_dump(named_vector, former_name = TRUE, collapse = ",")
}
\arguments{
\item{named_vector}{a named vector}

\item{former_name}{if TRUE, put names to former}

\item{collapse}{collapse separator}
}
\value{
character
}
\description{
dump a named vector into character
}
\examples{
vector_dump(c(e=1:4), former_name = TRUE,  collapse=';')
}
