% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{write_excel}
\alias{write_excel}
\title{write a tibble into an excel file}
\usage{
write_excel(df, filename, sheetname = "sheet1", creator = "")
}
\arguments{
\item{df}{tibble}

\item{filename}{the output filename}

\item{sheetname}{the sheet name, \code{sheet1} as default}

\item{creator}{creator}
}
\value{
return status
}
\description{
write a tibble into an excel file
}
\examples{
write_excel(mini_diamond, "mini_diamond.xlsx")
}
