% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{atomic_expr}
\alias{atomic_expr}
\title{whether the expression is an atomic one}
\usage{
atomic_expr(ex)
}
\arguments{
\item{ex}{expression}
}
\value{
logical value
}
\description{
whether the expression is an atomic one
}
\examples{
atomic_expr(rlang::expr(x))

atomic_expr(rlang::expr(!x))

atomic_expr(rlang::expr(x + y))

atomic_expr(rlang::expr(x > 1))

atomic_expr(rlang::expr(!x + y))

atomic_expr(rlang::expr(x > 1 | y < 2))

}
