% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baizer-package.R
\docType{package}
\name{baizer-package}
\alias{baizer}
\alias{baizer-package}
\title{baizer: Useful Functions for Data Processing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

In ancient Chinese mythology, Bai Ze is a divine creature that understands everything. `baizer` provides data processing functions frequently used by the author. Hope this package also knows what you want!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/william-swl/baizer}
  \item Report bugs at \url{https://github.com/william-swl/baizer/issues}
}

}
\author{
\strong{Maintainer}: William Song \email{william_swl@163.com}

}
\keyword{internal}
