% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{diff_index}
\alias{diff_index}
\title{the index of nth different character}
\usage{
diff_index(s1, s2, nth = 0)
}
\arguments{
\item{s1}{string1}

\item{s2}{string2}

\item{nth}{return the index of nth different character.
if \code{0} return all the indices}
}
\value{
the index of differences
}
\description{
the index of nth different character
}
\examples{
diff_index("ATTC", "ATAC")
}
