% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{fancy_count}
\alias{fancy_count}
\title{better count to show a main column and a fine column}
\usage{
fancy_count(df, main_group, fine_group, fine_fmt = "count", sort = TRUE)
}
\arguments{
\item{df}{tibble}

\item{main_group}{main column as main group}

\item{fine_group}{fine column as subgroup}

\item{fine_fmt}{output fine column format, \code{count|ratio|clean}}

\item{sort}{sort by frequency or not}
}
\value{
tibble
}
\description{
better count to show a main column and a fine column
}
\examples{
fancy_count(mini_diamond, "cut", "clarity")
}
