% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{fix_to_regex}
\alias{fix_to_regex}
\title{trans fixed string into regular expression string}
\usage{
fix_to_regex(p)
}
\arguments{
\item{p}{raw fixed pattern}
}
\value{
regex pattern
}
\description{
trans fixed string into regular expression string
}
\examples{
fix_to_regex("ABC|?(*)")
}
