% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{fps_vector}
\alias{fps_vector}
\title{farthest point sampling (FPS) for a vector}
\usage{
fps_vector(v, n, method = "round")
}
\arguments{
\item{v}{vector}

\item{n}{sample size}

\item{method}{\code{round|floor|ceiling}, the method used when trans to integer}
}
\value{
sampled vector
}
\description{
farthest point sampling (FPS) for a vector
}
\examples{
fps_vector(1:10, 4)
}
