% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{\%nin\%}
\alias{\%nin\%}
\title{not in calculation operator}
\usage{
left \%nin\% right
}
\arguments{
\item{left}{left element}

\item{right}{right element}
}
\value{
logical value, TRUE if left is not in right
}
\description{
not in calculation operator
}
\examples{
0 \%nin\% 1:4
}
