% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{hist_bins}
\alias{hist_bins}
\title{separate numeric x into bins}
\usage{
hist_bins(x, bins = 10, lim = c(min(x), max(x)), breaks = NULL, sort = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{bins}{bins number, defaults to 10}

\item{lim}{the min and max limits of bins, default as \code{c(min(x), max(x))}}

\item{breaks}{assign breaks directly and will ignore \code{bins} and \code{lim}}

\item{sort}{sort the result tibble}
}
\value{
tibble
}
\description{
separate numeric x into bins
}
\examples{

x <- dplyr::pull(mini_diamond, price, id)

hist_bins(x, bins = 20)

}
