% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-basic.R
\name{is.exist}
\alias{is.exist}
\title{if variable exists}
\usage{
is.exist(x, ls = NULL, envir = environment())
}
\arguments{
\item{x}{variable}

\item{ls}{variable list}

\item{envir}{environment}
}
\value{
logical value, TRUE if x exists
}
\description{
if variable exists
}
\examples{
is.exist(var_not_in_session)
}
