% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{percent_to_float}
\alias{percent_to_float}
\title{from percent number to float number}
\usage{
percent_to_float(x, digits = 2, to_double = FALSE)
}
\arguments{
\item{x}{percent number character}

\item{digits}{hold n digits after the decimal point}

\item{to_double}{use double output}
}
\value{
float character or double of x
}
\description{
from percent number to float number
}
\examples{
percent_to_float("12\%")
}
