% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{signif_round_string}
\alias{signif_round_string}
\title{signif or round strings}
\usage{
signif_round_string(x, digits = 2, format = "short")
}
\arguments{
\item{x}{number}

\item{digits}{signif or round digits}

\item{format}{short or long}
}
\value{
signif or round strings
}
\description{
signif or round strings
}
\examples{
signif_round_string(0.03851)
}
