% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_test}
\alias{stat_test}
\title{statistical test which returns a extensible tibble}
\usage{
stat_test(
  df,
  y,
  x,
  .by = NULL,
  trans = "identity",
  paired = FALSE,
  alternative = "two.sided",
  method = "wilcoxon",
  ns_symbol = "NS"
)
}
\arguments{
\item{df}{tibble}

\item{y}{value}

\item{x}{sample test group}

\item{.by}{super-group}

\item{trans}{scale transformation}

\item{paired}{paired samples or not}

\item{alternative}{one of "two.sided" (default), "greater" or "less"}

\item{method}{test method, 'wilcoxon' as default}

\item{ns_symbol}{symbol of nonsiginficant, 'NS' as default}
}
\value{
test result tibble
}
\description{
statistical test which returns a extensible tibble
}
\examples{
stat_test(mini_diamond, y = price, x = cut, .by = clarity)
}
